requirements.txt
requests
customtkinter
Pillow
packaging
